requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
alabaster
annotated-types==0.7.0
anyio==4.5.2
appdirs==1.4.4
argon2-cffi
argon2-cffi-bindings
arrow
astroid
astropy
asttokens
async-timeout==5.0.1
asyncer==0.0.8
atomicwrites==1.4.0
attrs==25.3.0
Automat
autopep8
Babel
backcall
backports.functools-lru-cache
backports.tempfile
backports.weakref==1.0.post1
bcrypt
beautifulsoup4
binaryornot
bitarray
bkcharts==0.2
black==19.10b0
bleach
bokeh
boto3
botocore
Bottleneck
brotlipy==0.7.0
cachetools
certifi==2025.4.26
cffi
chardet
charset-normalizer==3.4.2
click==8.1.8
cloudinary>=1.34.0
cloudpickle
colorama
colorcet
coloredlogs==15.0.1
constantly==15.1.0
cookiecutter
cryptography
cssselect==1.1.0
cycler==0.12.1
Cython
daal4py==2021.5.0
dask
datashader
debugpy
decorator
defusedxml
diff-match-patch
distributed
docutils
entrypoints
et-xmlfile==1.1.0
exceptiongroup==1.3.0
executing
fastapi==0.115.12
fastjsonschema
ffmpy==0.5.0
filelock==3.16.1
filetype==1.2.0
flake8
Flask
flatbuffers==25.2.10
fonttools==4.57.0
frozendict==2.4.2
frozenlist==1.5.0
fsspec==2025.3.0
future
gensim
glob2
gmpy2
google-api-core
google-auth
google-cloud-core
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
gradio==4.44.1
gradio_client==1.3.0
greenlet
grpcio
h11==0.16.0
hf-xet==1.1.2
h5py
HeapDict
holoviews
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.2
humanfriendly==10.0
hvplot
hyperlink
idna==3.10
imagecodecs
imageio==2.35.1
imagesize
importlib-metadata==8.5.0
importlib_resources==6.5.2
incremental
inflection==0.5.1
iniconfig
intake
intervaltree
ipykernel
ipython
ipython-genutils
ipywidgets
isort
itemadapter
itemloaders
itsdangerous
jdcal
jedi
Jinja2==3.1.6
jinja2-time
jmespath
joblib
json5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client
jupyter-core
jupyter-server
jupyterlab
jupyterlab-pygments
jupyterlab-server
jupyterlab-widgets
keyring
kiwisolver==1.4.7
lazy-object-proxy
lazy_loader==0.4
llvmlite==0.43.0
locket
lxml==5.2.1
Markdown
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline
mccabe==0.6.1
mdurl==0.1.2
mistune
mock
mpmath==1.3.0
msgpack
multidict==6.1.0
multipledispatch
multitasking==0.0.11
munkres==1.1.4
mypy-extensions==0.4.3
nbclassic
nbclient
nbconvert
nbformat
nest-asyncio
networkx==3.1
nltk
nose
notebook
numba==0.58.1
numexpr
numpy==1.26.4
numpydoc
olefile
onnxruntime==1.19.2
opencv-python-headless==4.11.0.86
openpyxl
orjson==3.10.18
packaging
panel
param
parsel
parso
partd
pathlib-abc==0.1.1
pathspec==0.7.0
patsy==0.5.2
pandas==2.0.3
peewee==3.17.1
pep8==1.7.1
pexpect
pickleshare
Pillow==10.4.0
pkginfo
platformdirs==4.3.6
plotly
pluggy
pooch==1.8.2
poyo
prometheus-client
prompt-toolkit
protobuf==5.29.4
psutil
ptyprocess
pure-eval
pyasn1==0.2.8
pyasn1-modules==0.2.8
pycodestyle
pycosat==0.6.3
pycparser
pyct
pycurl
pydantic==2.11.5
pydantic_core==2.33.2
PyDispatcher==2.0.5
pydocstyle
pydub==0.25.1
pyerfa
pyflakes
Pygments==2.19.1
PyHamcrest
PyJWT
pylint
pyls-spyder==0.4.0
PyMatting==1.1.14
pyparsing==3.1.4
pyrsistent
python-dateutil==2.9.0.post0
python-lsp-black
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
python-multipart==0.0.20
python-slugify
python-snappy
pytz==2025.2
pyviz-comms
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq
queuelib==1.5.0
rafds-py==0.20.1
regex==2024.11.6
rembg==2.0.61
requests==2.32.3
requests-file
rich==14.0.0
rope
rsa
Rtree
ruff==0.11.11
s3transfer
scikit-image==0.21.0
scikit-learn
scikit-learn-intelex==2021.20220215.132722
scipy==1.10.1
Scrapy
seaborn
semantic-version==2.10.0
Send2Trash
service-identity
shellingham==1.5.4
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
snowballstemmer
sortedcollections
sortedcontainers
soupsieve
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
SQLAlchemy
stack-data
starlette==0.46.2
statsmodels
sympy
tables
tabulate==0.8.9
tblib
tenacity
terminado
testpath
text-unidecode
textdistance
threadpoolctl
three-merge
tifffile==2023.7.10
tinycss
tldextract
toml
tomli
tomlkit==0.12.0
toolz
tornado
traitlets
Twisted
typed-ast==0.4.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ujson
Unidecode
urllib3==2.4.0
uvicorn==0.34.2
watchdog==4.0.2
webencodings==0.5.1
websocket-client
websockets==12.0
Werkzeug
widgetsnbextension
wrapt
wurlitzer
xarray
xlrd
XlsxWriter
yapf
yarl==1.15.2
yfinance==0.2.41
zict==2.0.0
zipp==3.20.2
zope.interface
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
